requirements.txt
click==6.7
jsonschema
requests==2.18.4
scout.py==0.3.0
pathlib2==2.3.0
